Time: 0 ms (100%), Space: 113.9 MB (27.95%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -17,7 +17,7 @@ public:
             right++;
             sum+=nums[right];
 
-            double avg = (double)sum/k;
+            double avg =sum/k;
             ans = max(ans, avg);
         }
         return ans;
